Update changelog for version 2.6.1, optimize network packet handling, and add Gemini attribution to README

+++ Crossplay/PacketFactory.cs
@@ -21,7 +21,7 @@ namespace Crossplay
         {
             long currentPosition = writer.BaseStream.Position;
             writer.BaseStream.Position = 2L;
-            writer.Write(type);
+            writer.Write((byte)type);
             writer.BaseStream.Position = currentPosition;
             return this;
         }
